+++ R/operator.R
@@ -1,30 +1,30 @@
-#' Unpacking operator
+#' Multiple assignment operators
 #'
 #' Assign values to name(s).
 #'
-#' @param x A name structure, see details.
+#' @param x A name structure, see section below.
 #'
 #' @param value A list of values, vector of values, or \R object to assign.
 #'
-#' @section Left-hand side syntax:
+#' @section Name Structure:
 #'
 #' **the basics**
 #'
-#' At its simplest the left-hand side may be a single variable name, in which
-#' case \code{\%<-\%} performs regular assignment,
-#' \code{x \%<-\% list(1, 2, 3)}.
+#' At its simplest, the name structure may be a single variable name, in which
+#' case \code{\%<-\%} and \code{\%->\%} perform regular assignment, \code{x
+#' \%<-\% list(1, 2, 3)} or \code{list(1, 2, 3) \%->\% x}.
 #'
 #' To specify multiple variable names use a call to `c()`, for example
 #' \code{c(x, y, z) \%<-\% c(1, 2, 3)}.
 #'
-#' When `value` is neither an atomic vector nor a list, \code{\%<-\%} will try
-#' to destructure `value` into a list before assigning variables, see
-#' [destructure()].
+#' When `value` is neither an atomic vector nor a list, \code{\%<-\%} and
+#' \code{\%->\%} will try to destructure `value` into a list before assigning
+#' variables, see [destructure()].
 #'
 #' **nested names**
 #'
 #' One can also nest calls to `c()` when needed, `c(x, c(y, z))`. This nested
-#' structure is used to unpack nested right-hand side values,
+#' structure is used to unpack nested values,
 #' \code{c(x, c(y, z)) \%<-\% list(1, list(2, 3))}.
 #'
 #' **collector variables**
@@ -41,12 +41,18 @@
 #' Use `...` to skip multiple values without raising an error or assigning the
 #' values, \code{c(w, ..., z) \%<-\% list(1, NA, NA, 4)}.
 #'
+#' **default values**
+#'
+#' Use `=` to specify a default value for a variable, \code{c(x, y = NULL)
+#' \%<-\% tail(1, 2)}.
+#'
 #' @return
 #'
-#' \code{\%<-\%} invisibly returns `value`.
+#' \code{\%<-\%} and \code{\%->\%} invisibly return `value`.
 #'
-#' \code{\%<-\%} is used primarily for its assignment side-effect. \code{\%<-\%}
-#' assigns into the environment in which it is evaluated.
+#' These operators are used primarily for their assignment side-effect.
+#' \code{\%<-\%} and \code{\%->\%} assign into the environment in which they
+#' are evaluated.
 #'
 #' @seealso
 #'
@@ -54,7 +60,6 @@
 #' [destructure()].
 #'
 #' @name operator
-#' @md
 #' @export
 #' @examples
 #' # basic usage
@@ -147,46 +152,134 @@
 #' firsts  # c("Nathan", "Maria", ..
 #' lasts   # c("Smith", "Peterson", ..
 #'
+#' # handle missing values with default values
+#' parse_time <- function(x) {
+#'   strsplit(x, " ")[[1]]
+#' }
+#'
+#' c(hour, period = NA) %<-% parse_time("10:00 AM")
+#'
+#' hour    # "10:00"
+#' period  # "AM"
+#'
+#' c(hour, period = NA) %<-% parse_time("15:00")
+#'
+#' hour    # "15:00"
+#' period  # NA
+#'
+#' # right operator
+#' list(1, 2, "a", "b", "c") %->% c(x, y, ...chars)
+#'
+#' x      # 1
+#' y      # 2
+#' chars  # list("a", "b", "c")
+#'
+#' # magrittr chains, install.packages("magrittr") for this example
+#' if (requireNamespace("magrittr", quietly = TRUE)) {
+#'   library(magrittr)
+#'
+#'   c("hello", "world!") %>%
+#'     paste0("\n") %>%
+#'     lapply(toupper) %->%
+#'     c(greeting, subject)
+#'
+#'   greeting  # "HELLO\n"
+#'   subject   # "WORLD!\n"
+#' }
+#'
 `%<-%` <- function(x, value) {
-  ast <- tree(substitute(x))
-  cenv <- parent.frame()
-
-  if (length(ast) != 1 && ast[[1]] != "c") {
-    return(old_operator(ast, value, cenv))
-  }
-
-  internals <- calls(ast)
-  lhs <- tryCatch(
-    variables(ast),
-    error = function(e) {
-      stop(
-        "invalid `%<-%` left-hand side, expecting symbol, but ", e$message,
-        call. = FALSE
-      )
+  tryCatch(
+    multi_assign(substitute(x), value, parent.frame()),
+    invalid_lhs = function(e) {
+      stop("invalid `%<-%` left-hand side, ", e$message, call. = FALSE)
+    },
+    invalid_rhs = function(e) {
+      stop("invalid `%<-%` right-hand side, ", e$message, call. = FALSE)
     }
   )
+}
+
+#' @rdname operator
+#' @export
+`%->%` <- function(value, x) {
+  tryCatch(
+    multi_assign(substitute(x), value, parent.frame()),
+    invalid_lhs = function(e) {
+      stop("invalid `%->%` right-hand side, ", e$message, call. = FALSE)
+    },
+    invalid_rhs = function(e) {
+      stop("invalid `%->%` left-hand side, ", e$message, call. = FALSE)
+    }
+  )
+}
+
+# The real power behind %->% and %<-%
+#
+# Within the function `lhs` and `rhs` refer to the left- and right-hand side of
+# a call to %<-% operator. For %->% the lhs and rhs from the original call are
+# swapped when passed to `multi_assign`.
+#
+# @param x A name structure, converted into a tree-like structure with `tree`.
+#
+# @param value The values to assign.
+#
+# @param env The environment where the variables will be assigned.
+#
+multi_assign <- function(x, value, env) {
+  ast <- tree(x)
+  internals <- calls(ast)
+  lhs <- variables(ast)
+  rhs <- value
 
   #
-  # standard assignment
+  # standard assignment, no calls (i.e. `c`) found
   #
   if (is.null(internals)) {
-    assign(as.character(ast), value, envir = cenv)
+    assign(as.character(ast), value, envir = env)
     return(invisible(value))
   }
 
-  if (any(internals != "c")) {
-    name <- internals[which(internals != "c")][1]
-    stop(
-      "invalid `%<-%` left-hand side, unexpected call `", name, "`",
-      call. = FALSE
-    )
+  #
+  # *error* multiple assignment, but sinle RHS value
+  #
+  if (length(value) == 0) {
+    stop_invalid_rhs(incorrect_number_of_values())
   }
 
-  # if (is_list(lhs) && is_list(car(lhs))) {
-  #   lhs <- car(lhs)
-  # }
+  #
+  # edge cases when RHS is not a list
+  #
+  if (!is_list(value)) {
+    if (is.atomic(value)) {
+      rhs <- as.list(value)
+    } else {
+      rhs <- destructure(value)
+    }
+  }
 
-  massign(lhs, value, envir = cenv)
+  #
+  # tuples in question are variable names and value to assign
+  #
+  tuples <- pair_off(lhs, rhs, env)
+
+  for (t in tuples) {
+    name <- t[["name"]]
+    val <- t[["value"]]
+
+    #
+    # collector variable names retain the leading "..." in order to revert
+    # list values back to vectors if necessary
+    #
+    if (is_collector(name)) {
+      name <- sub("^\\.\\.\\.", "", name)
+
+      if (is.atomic(value)) {
+        val <- unlist(val)
+      }
+    }
+
+    assign(name, val, envir = env)
+  }
 
   invisible(value)
 }