+++ tests/testthat/test-operator.R
@@ -101,6 +101,33 @@ test_that("%<-% skips multiple values using ...", {
   expect_equal(b, 4)
 })
 
+test_that("%<-% assigns default values", {
+  c(a, b = 1) %<-% c(3)
+  expect_equal(a, 3)
+  expect_equal(b, 1)
+
+  c(d, e = iris, f = 3030) %<-% 5
+  expect_equal(d, 5)
+  expect_equal(e, iris)
+  expect_equal(f, 3030)
+})
+
+test_that("%<-% assign default value of NULL", {
+  c(a, b = NULL) %<-% c(3)
+  expect_equal(a, 3)
+  expect_equal(b, NULL)
+})
+
+test_that("%<-% default values do not override specified values", {
+  c(a = 1, b = 4) %<-% 2
+  expect_equal(a, 2)
+  expect_equal(b, 4)
+
+  c(d = 5, e = 6) %<-% c(8, 9)
+  expect_equal(d, 8)
+  expect_equal(e, 9)
+})
+
 test_that("%<-% throws error on unequal number of variables and values", {
   expect_error(
     c(a, b) %<-% list(1),
@@ -136,20 +163,27 @@ test_that("%<-% throws error when invalid calls on LHS", {
   )
 })
 
+test_that("%<-% throws error when blank variable names", {
+  expect_error(
+    c(, a) %<-% c(1, 2),
+    "invalid `%<-%` left-hand side, found empty variable, check for extraneous commas"
+  )
+})
+
 test_that('%<-% throws error when invalid "variables" on LHS', {
   expect_error(
     c(mean(1, 2), a) %<-% list(1, 2),
-    "invalid `%<-%` left-hand side, expecting symbol, but found numeric"
+    "invalid `%<-%` left-hand side, unexpected call `mean`"
   )
 
   expect_error(
     c(a, f()) %<-% list(1, 2),
-    "invalid `%<-%` left-hand side, expecting symbol, but found call"
+    "invalid `%<-%` left-hand side, expected symbol, but found call"
   )
 
   expect_error(
     f() %<-% list(1),
-    "invalid `%<-%` left-hand side, expecting symbol, but found call"
+    "invalid `%<-%` left-hand side, expected symbol, but found call"
   )
 })
 
@@ -157,5 +191,25 @@ test_that("%<-% throws error when assigning empty list", {
   expect_error(
     c(a, b) %<-% list(),
     "invalid `%<-%` right-hand side, incorrect number of values"
+  )
+})
+
+test_that("%->% properly calls `multi_assign`", {
+  c(1, 2) %->% c(x, y)
+  expect_equal(x, 1)
+  expect_equal(y, 2)
+})
+
+test_that("%->% errors include %->% in message, flips lhs and rhs", {
+  expect_error(
+    c(1, 2) %->% {x:y},
+    "invalid `%->%` right-hand side, unexpected call `{`",
+    fixed = TRUE
+  )
+
+  expect_error(
+    1 %->% c(x, y),
+    "invalid `%->%` left-hand side, incorrect number of values",
+    fixed = TRUE
   )
 })